Re-enable Lambda update; disable Serilog self-logging
Re-enabled the step in `main.yaml` to update the Development Lambda
function when the GitHub reference is the 'development' branch.
This step was previously commented out.

Commented out the line in `Program.cs` that enables Serilog's
self-logging to a file, stopping internal debug and error messages
from being written to `serilog-selflog.txt`.